Check for bcc_email

+++ database/migrations/2017_01_01_214241_add_inclusive_taxes.php
@@ -59,7 +59,9 @@ class AddInclusiveTaxes extends Migration
 
         Schema::table('accounts', function ($table) {
             $table->enum('bluevine_status', ['ignored', 'signed_up'])->nullable();
-            $table->dropColumn('bcc_email');
+            if (Schema::hasColumn('accounts', 'bcc_email')) {
+                $table->dropColumn('bcc_email');
+            }
             $table->dropColumn('client_number_prefix');
             $table->dropColumn('client_number_counter');
             $table->dropColumn('client_number_pattern');
